Scorpion graph- check if graph is a scorpion
A scorpion is an undirected graph G of the following form: there are three
special vertices, called the sting, the tail, and the body, of degree 1, 2
and n -2 respectively. The sting is connected only to the tail the tail is
connected only to the sting and the body and the body is connected to
all vertices except the sting. The other vertices of G may be connected to
each other arbit